#1 Image file upload now functional
The result of the /:heretakethis uri (to be changed to something more
sensible, later) is the very image passed in.  This will need some
reworking between the server and client side of this application, at to
whether the ":heretakethis" call returns the image, or should return
only an OK, with the image extracted at the client side's leisure.
[fixes #1]